requirements.txt
anyascii==0.3.0
asgiref==3.4.1
beautifulsoup4==4.9.3
certifi==2021.10.8
charset-normalizer==2.0.7
dj-database-url==0.5.0
Django==3.2.8
django-address==0.2.5
django-cogwheels==0.3
django-credit-cards==0.4.1
django-filter==2.4.0
django-heroku==0.3.1
django-modelcluster==5.2
django-multiselectfield==0.1.12
django-taggit==1.5.1
django-treebeard==4.5.1
djangorestframework==3.12.4
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
gunicorn==20.1.0
html5lib==1.1
idna==3.3
l18n==2020.6.1
openpyxl==3.0.9
Pillow==8.4.0
psycopg2==2.9.1
pytz==2021.3
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
tablib==3.1.0
telepath==0.2
Unidecode==1.3.2
urllib3==1.26.7
wagtail==2.14.2
wagtailmenus==3.0.2
wagtailstreamforms==3.16.3
webencodings==0.5.1
whitenoise==5.3.0
Willow==1.4
xlrd==2.0.1
XlsxWriter==1.4.5
xlwt==1.3.0
